analysis: Add WHO data to processing pipeline
I don't want to merge cases into the flights right now because I'm not
sure there's a single way I want to do it for this analysis. Do we
include all cases worldwide? The sum of cases between origin and
destination countries? We may have to try multiple ones in the final
data analysis, so I'll do it there.