analysis: Fix misinterpretation of data
These are not just US domestic and international flights: they're
whatever the US government collected. (I wouldn't be surprised if they
had some international network of ADS-B receivers or something.) In any
case, that was a mistake.

I also tried to explain the dip and spike in the COVID data more
clearly.